requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2
django-crispy-forms==1.11.2
django-session-timeout==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.4.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
